Move all parsing to language_parser.rs
