+++ src/untyped_language/ast.rs
@@ -1,6 +1,6 @@
 use crate::parser_combinator::Token;
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum Value {
     Number(i32),
     Bool(bool),
@@ -9,7 +9,7 @@ pub enum Value {
 #[derive(Debug, Clone)]
 pub struct Parameter(pub String, pub String); //name, type
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct Fun {
     pub name: Token<String>,
     pub params: Vec<Token<Parameter>>,
